removed old integration tests
