+++ compose.yml
@@ -1,6 +1,6 @@
 services:
   localstack:
-    image: localstack/localstack:3.0.2
+    image: localstack/localstack
     ports:
       - '4566:4566' # LocalStack Gateway
       - '4510-4559:4510-4559' # external services port range
@@ -9,8 +9,9 @@ services:
     environment:
       DEBUG: ${DEBUG:-1}
       LS_LOG: WARN # Localstack DEBUG Level
-      SERVICES: s3,sqs,sns,firehose
+      SERVICES: sqs
       LOCALSTACK_HOST: 127.0.0.1
+      TZ: Europe/London
     volumes:
       - '${TMPDIR:-/tmp}/localstack:/var/lib/localstack'
       - ./compose/start-localstack.sh:/etc/localstack/init/ready.d/start-localstack.sh
@@ -20,48 +21,7 @@ services:
       start_period: 5s
       retries: 10
 
-  # redis:
-  #   image: redis:7.2.3-alpine3.18
-  #   ports:
-  #     - '6379:6379'
-  #   restart: always
-  #   networks:
-  #     - cdp-tenant
-
-  # mongodb:
-  #   image: mongo:6.0.13
-  #   networks:
-  #     - cdp-tenant
-  #   ports:
-  #     - '27017:27017'
-  #   volumes:
-  #     - mongodb-data:/data
-  #   restart: always
-
   ################################################################################
-
-  # your-frontend:
-  #   image: defradigital/your-frontend:${YOUR_FRONTEND_VERSION:-latest}
-  #   ports:
-  #     - '3000:3000'
-  #   links:
-  #     - 'localstack:localstack'
-  #     - 'redis:redis'
-  #   depends_on:
-  #     localstack:
-  #       condition: service_healthy
-  #     redis:
-  #       condition: service_started
-  #   env_file:
-  #     - 'compose/aws.env'
-  #   environment:
-  #     PORT: 3000
-  #     NODE_ENV: development
-  #     REDIS_HOST: redis
-  #     LOCALSTACK_ENDPOINT: http://localstack:4566
-  #     USE_SINGLE_INSTANCE_CACHE: true
-  #   networks:
-  #     - cdp-tenant
 
   decision-deriver:
     build:
@@ -73,12 +33,22 @@ services:
     depends_on:
       localstack:
         condition: service_healthy
+      wiremock:
+        condition: service_started
     env_file:
       - 'compose/aws.env'
     environment:
       ASPNETCORE_ENVIRONMENT: Development
+      DataApi__BaseAddress: "http://wiremock"
       PORT: 8080
       SQS_Endpoint: http://sqs.eu-west-2.localhost.localstack.cloud:4566
+  
+  wiremock:
+    environment:
+      TZ: Europe/London
+    image: sheyenrath/wiremock.net:latest
+    ports:
+      - "9090:80"
     
 
 ################################################################################
