+++ tests/Deriver.IntegrationTests/Consumers/CustomsDeclarationsConsumerTests.cs
@@ -1,8 +1,5 @@
 using System.Net;
-using System.Security.Cryptography;
 using System.Text.Json;
-using Amazon.Runtime;
-using Amazon.SQS;
 using Amazon.SQS.Model;
 using Defra.TradeImportsDataApi.Domain.Events;
 using Defra.TradeImportsDecisionDeriver.TestFixtures;
@@ -12,7 +9,7 @@ using WireMock.Client;
 using WireMock.Client.Extensions;
 using Xunit.Abstractions;
 
-namespace Defra.TradeImportsDecisionDeriver.Deriver.IntegrationTests;
+namespace Defra.TradeImportsDecisionDeriver.Deriver.IntegrationTests.Consumers;
 
 public class CustomsDeclarationsConsumerTests(ITestOutputHelper output) : SqsTestBase(output)
 {
@@ -35,9 +32,6 @@ public class CustomsDeclarationsConsumerTests(ITestOutputHelper output) : SqsTes
         var importNotification = ImportPreNotificationFixtures.ImportPreNotificationCreatedFixture();
 
         var customsDeclaration = CustomsDeclarationResponseFixtures.CustomsDeclarationResponseFixture();
-
-        ////var mrn = GenerateMrn();
-        ////var clearanceRequest = ClearanceRequestFixture(mrn).Create();
 
         await _wireMockAdminApi.ResetMappingsAsync();
         await _wireMockAdminApi.ResetRequestsAsync();
@@ -81,47 +75,5 @@ public class CustomsDeclarationsConsumerTests(ITestOutputHelper output) : SqsTes
                 return requests.Count == 1;
             })
         );
-    }
-}
-
-public class SqsTestBase(ITestOutputHelper output)
-{
-    private const string QueueUrl =
-        "http://sqs.eu-west-2.127.0.0.1:4566/000000000000/trade_imports_data_import_declaration_upserts";
-
-    private readonly AmazonSQSClient _sqsClient = new(
-        new BasicAWSCredentials("test", "test"),
-        new AmazonSQSConfig { AuthenticationRegion = "eu-west-2", ServiceURL = "http://localhost:4566" }
-    );
-
-    protected Task<ReceiveMessageResponse> ReceiveMessage()
-    {
-        return _sqsClient.ReceiveMessageAsync(QueueUrl, CancellationToken.None);
-    }
-
-    protected Task<GetQueueAttributesResponse> GetQueueAttributes()
-    {
-        return _sqsClient.GetQueueAttributesAsync(
-            new GetQueueAttributesRequest { AttributeNames = ["ApproximateNumberOfMessages"], QueueUrl = QueueUrl },
-            CancellationToken.None
-        );
-    }
-
-    protected async Task SendMessage(
-        string messageGroupId,
-        string body,
-        Dictionary<string, MessageAttributeValue>? messageAttributes = null
-    )
-    {
-        var request = new SendMessageRequest
-        {
-            MessageAttributes = messageAttributes,
-            MessageBody = body,
-            ////MessageDeduplicationId = RandomNumberGenerator.GetString("abcdefg", 20),
-            ////MessageGroupId = messageGroupId,
-            QueueUrl = QueueUrl,
-        };
-        var result = await _sqsClient.SendMessageAsync(request, CancellationToken.None);
-        output.WriteLine("Sent {0} to {1}", result.MessageId, QueueUrl);
     }
 }
