--- tests/Deriver.IntegrationTests/Consumers/ImportPreNotificationConsumerTests.cs
@@ -1,55 +1,0 @@
-using Defra.TradeImportsDataApi.Api.Client;
-using Defra.TradeImportsDataApi.Domain.CustomsDeclaration;
-using Defra.TradeImportsDataApi.Domain.Events;
-using Defra.TradeImportsDecisionDeriver.TestFixtures;
-using NSubstitute;
-using Xunit.Abstractions;
-
-namespace Defra.TradeImportsDecisionDeriver.Deriver.IntegrationTests.Consumers;
-
-[Collection("Non-Parallel Collection")]
-public class ImportPreNotificationConsumerTests(DeriverWebApplicationFactory factory, ITestOutputHelper outputHelper)
-    : IntegrationTests(factory, outputHelper)
-{
-    [Fact]
-    public async Task EndToEndTest()
-    {
-        var importNotification = ImportPreNotificationFixtures.ImportPreNotificationCreatedFixture();
-
-        var customsDeclaration = CustomsDeclarationResponseFixtures.CustomsDeclarationResponseFixture();
-        _apiClient
-            .GetCustomsDeclarationsByChedId(importNotification.ResourceId, Arg.Any<CancellationToken>())
-            .Returns([customsDeclaration]);
-
-        _apiClient
-            .GetImportPreNotificationsByMrn(customsDeclaration.MovementReferenceNumber, Arg.Any<CancellationToken>())
-            .Returns(
-                [
-                    new ImportPreNotificationResponse(
-                        ImportPreNotificationFixtures.ImportPreNotificationFixture("test")!,
-                        DateTime.Now,
-                        DateTime.Now
-                    ),
-                ]
-            );
-
-        _apiClient
-            .GetCustomsDeclaration(customsDeclaration.MovementReferenceNumber, Arg.Any<CancellationToken>())
-            .Returns(customsDeclaration);
-
-        await PurgeQueue();
-        await SendMessage(importNotification, ResourceEventResourceTypes.ImportPreNotification);
-
-        await WaitOnQueueBeingProcessed();
-
-        await _apiClient
-            .Received(1)
-            .PutCustomsDeclaration(
-                Arg.Any<string>(),
-                Arg.Any<CustomsDeclaration>(),
-                Arg.Any<string>(),
-                Arg.Any<CancellationToken>()
-            );
-        await _factory.DisposeAsync();
-    }
-}
